+++ collector/exporter/otelarrowexporter/internal/arrow/bestofn.go
@@ -8,6 +8,7 @@ import (
 	"math/rand"
 	"runtime"
 	"sort"
+	"time"
 )
 
 // bestOfNPrioritizer is a prioritizer that selects a less-loaded stream to write.
@@ -42,7 +43,7 @@ type streamSorter struct {
 
 var _ streamPrioritizer = &bestOfNPrioritizer{}
 
-func newBestOfNPrioritizer(dc doneCancel, numChoices, numStreams int, lf loadFunc) (*bestOfNPrioritizer, []*streamWorkState) {
+func newBestOfNPrioritizer(dc doneCancel, numChoices, numStreams int, lf loadFunc, maxLifetime time.Duration) (*bestOfNPrioritizer, []*streamWorkState) {
 	var state []*streamWorkState
 
 	// Limit numChoices to the number of streams.
@@ -50,8 +51,9 @@ func newBestOfNPrioritizer(dc doneCancel, numChoices, numStreams int, lf loadFun
 
 	for i := 0; i < numStreams; i++ {
 		ws := &streamWorkState{
-			waiters: map[int64]chan<- error{},
-			toWrite: make(chan writeItem, 1),
+			maxStreamLifetime: addJitter(maxLifetime),
+			waiters:           map[int64]chan<- error{},
+			toWrite:           make(chan writeItem, 1),
 		}
 
 		state = append(state, ws)